requirements.txt
streamlit==1.29.0
requests==2.32.2
python-dotenv==1.0.0
pandas==2.1.4
plotly==5.18.0
streamlit-lottie==0.0.5
pillow==10.1.0